tangle - extract `platform/action` (fix #141154)
