+++ src/vs/workbench/contrib/issue/electron-sandbox/issue.contribution.ts
@@ -5,7 +5,8 @@
 
 import { localize } from 'vs/nls';
 import product from 'vs/platform/product/common/product';
-import { ICommandAction, MenuRegistry, MenuId, registerAction2 } from 'vs/platform/actions/common/actions';
+import { MenuRegistry, MenuId, registerAction2 } from 'vs/platform/actions/common/actions';
+import { ICommandAction } from 'vs/platform/action/common/action';
 import { CATEGORIES } from 'vs/workbench/common/actions';
 import { ReportPerformanceIssueUsingReporterAction, OpenProcessExplorer } from 'vs/workbench/contrib/issue/electron-sandbox/issueActions';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
